Created the initial parts of the AutoLineup and Fixed AtuoLineup
Added the drivetrain requires to AutoLineup. This is so the future auto lineup can take control of the drivetrain.

Edited the talon instantiation to use the Talon factory. This is so certain cantalon settings are set for the talons. Furthermore i edited the elevator talons to avoid the situation where they run against each other.