Added some things for climber testing PEOPLE TESTING READ THIS BEFORE
Specifically added some data points on the smart dashboard including being able to see the difference between the two posts. I also added the climber demand functionality to some joystick buttons for easy testing, What has to be done tonight is adding the PID loop logic to the climber as well as the climb demands. I see that this is currently missing and needs to be done.